changed from int to byte

+++ cpu.h
@@ -59,7 +59,7 @@ typedef struct Chip8
     //we good?
     byte status;
     
-    int drawFlag;
+    byte drawFlag;
     
 } Chip8;
